requirements.txt
flask
requests
flasgger
python-dotenv
PyJWT